% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{ldap_bind}
\alias{ldap_bind}
\title{Simple bind}
\usage{
ldap_bind(self, private, user, pw, type = c("cn", "uid"), timeout)
}
\arguments{
\item{self}{Self from R6 object}

\item{private}{Private list from R6 object}

\item{user}{User to authenticate}

\item{pw}{Password to authenticate}

\item{type}{Type of object to authenticate with - one of either `cn` or `uid`}

\item{timeout}{Timeout for authentication}
}
\description{
Perform a simple LDAP bind operation
}
\keyword{internal}
