% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{ldap_reset}
\alias{ldap_reset}
\title{Reset LDAP authentication}
\usage{
ldap_reset(private, self)
}
\arguments{
\item{private}{Private list from R6 object}

\item{self}{Self from R6 object}
}
\description{
Reset LDAP authentication
}
\keyword{internal}
