% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldapr.R
\docType{package}
\name{ldapr-package}
\alias{ldapr}
\alias{ldapr-package}
\title{ldapr: An R implementation of LDAP}
\description{
An implementation of LDAP in R. This allows authentication against a valid LDAP server, which is mostly useful inside a 
  shiny application that requires use authentication.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/liamgilbey/ldapr}
  \item Report bugs at \url{https://github.com/liamgilbey/ldapr/issues}
}

}
\author{
\strong{Maintainer}: Liam Gilbey \email{liam.gilbey@gmail.com}

}
\keyword{internal}
