% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ldapr_init}
\alias{ldapr_init}
\title{Initialize an LDAP connection handle}
\usage{
ldapr_init(ldap_uri)
}
\arguments{
\item{ldap_uri}{A valid LDAP uri string}
}
\description{
Initialize an LDAP connection handle
}
\keyword{internal}
