% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ldapr_is_ldap_url}
\alias{ldapr_is_ldap_url}
\title{Validate an LDAP URI. This is a fairly simple version of validation.}
\usage{
ldapr_is_ldap_url(ldap_uri)
}
\arguments{
\item{ldap_uri}{The URI you wish to validate}
}
\description{
Validate an LDAP URI. This is a fairly simple version of validation.
}
