% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ldapr_url_parse}
\alias{ldapr_url_parse}
\title{Validate an LDAP URI. This is a robust validation of an LDAP URI}
\usage{
ldapr_url_parse(ldap_uri)
}
\arguments{
\item{ldap_uri}{The URI you wish to validate}
}
\description{
Validate an LDAP URI. This is a robust validation of an LDAP URI
}
