% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldapr.R
\name{ldapsearch}
\alias{ldapsearch}
\title{Search LDAP}
\usage{
ldapsearch(...)
}
\arguments{
\item{...}{Parameters for the call to ldapsearch. Each parameter must be named
(e.g. h for the host definition) and may also be NULL for an parameter
without value. Some values, like D must be double quoted for evaluation within
quotes.}
}
\value{
Named list of returned attributes or NULL if the call wasn't successful.
}
\description{
This function serves as interface to the console function ldapsearch.
It allows to bind to a LDAP server and search for attributes.
The syntax of this function is based on the syntax of ldapsearch
\link{http://www.openldap.org/software/man.cgi?query=ldapsearch&sektion=1&manpath=OpenLDAP+2.4-Release}
}
\details{
This package requires the library openldap to be installed (https://www.openldap.org/).
}
\examples{
ldapsearch(x = NULL, h = 'ldap.myhost.org', D = '"QueryUserName\\@myDomain"', w = 'QueryUserPassword', b = '"CN=SomeUser,OU=Users,DC=SomeDomain"')

}
