% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ldapr_is_ldap_url}
\alias{ldapr_is_ldap_url}
\title{Validate an LDAP URL.}
\usage{
ldapr_is_ldap_url(ldap_uri)
}
\arguments{
\item{ldap_uri}{The URI you wish to validate}
}
\description{
This is a fairly simple version of validation, simply checking whether the URL string starts with \code{ldap://}
}
\keyword{internal}
