% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ldapr_url_parse}
\alias{ldapr_url_parse}
\title{Parse an LDAP URL.}
\usage{
ldapr_url_parse(ldap_uri)
}
\arguments{
\item{ldap_uri}{The URL to validate}
}
\description{
This function performs robust URL validation by attempting to break the URL into the component pieces.
}
\keyword{internal}
